requirements.txt
customtkinter==5.2.1
dbus-python==1.3.2
pillow<10.0.0
psutil==5.9.6
pybluez2==0.46
pystray==0.19.4
packaging